+++ src/clients/inworld.client.ts
@@ -137,6 +137,8 @@ export class InworldClient {
       .setAudio(capabilities?.audio ?? true)
       .setEmotions(capabilities?.emotions ?? false)
       .setInterruptions(capabilities?.interruptions ?? false)
+      .setPhonemeInfo(capabilities?.phonemes ?? false)
+      .setSilenceEvents(capabilities?.silence ?? false)
       .setText(true)
       .setTriggers(true);
   }