+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* Add phonemes event support
+* Add silence event support
+
 ## 2022-04-06 v1.1.8
 
 * Remove deprecated emotions attributes: joy, fear, trust and surprise